third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/106-import-xhtml-expected.txt

This is a testharness.js-based test.
FAIL Stylesheet in XHTML HEAD with @import blocking scripts assert_equals: expected "fixed" but got "static"
Harness: the test ran to completion.
